moving LocationExtension static class into same .cs file as Location class
